pyproject.toml
[project]
name = "unisearch"
version = "0.1.0"
description = "University search and comparison app with AI-powered recommendations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.3",
    "groq>=1.1.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-flask>=1.3.0",
    "pytest-mock>=3.14.0",
]
scrapers = [
    "aiohttp>=3.11.18",
    "openpyxl>=3.1.5",
]

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "migrations"]

[tool.pytest.ini_options]
testpaths = ["tests"]
